Small change preserved Python v2.6 compatibility; use subprocess instead of os.system.
With the `subprocess` module, you don't need to worry about having to quote a file name
with spaces.

E.g. with file '/home/carl/temp/My Obnoxious File With Spaces.txt'

Before:
  # Notice the double quoting required.
  ./when-changed -r /home/carl/temp -c /bin/ls -l '"%f"'

After:
  # More intuitive.
  ./when-changed -r /home/carl/temp -c /bin/ls -l %f

Also, at the recommendation of the Standard Library, use the drop-in replacement
`subprocess32` if it is available.